+++ packages/environment-check/test/checks/get-installed.test.ts
@@ -8,6 +8,8 @@ import {
 import fs from 'fs';
 import { isAppStudio } from '@sap-ux/btp-utils';
 import { getLogger } from '../../src/logger';
+import type { Extension } from 'vscode';
+import { t } from '../../src/i18n';
 
 jest.mock('@sap-ux/btp-utils', () => ({
     isAppStudio: jest.fn()
@@ -15,6 +17,9 @@ jest.mock('@sap-ux/btp-utils', () => ({
 const mockIsAppStudio = isAppStudio as jest.Mock;
 
 describe('Test install functions', () => {
+    const mockPackageJson =
+        '{  "name": "@sap/generator-fiori",  "displayName": "SAP Fiori application",  "version": "1.7.5",  "description": "Create an SAPUI5 application using SAP Fiori elements or a freestyle approach"  }';
+
     beforeEach(() => {
         jest.resetAllMocks();
     });
@@ -70,12 +75,48 @@ describe('Test install functions', () => {
         jest.spyOn(fs, 'readdirSync').mockImplementation(() => {
             throw new Error('Could not read directory');
         });
-        const result = await getInstalledExtensions(logger);
+        const result = await getInstalledExtensions(undefined, logger);
         const messages = logger.getMessages();
         expect(result).toBe(undefined);
         expect(messages).toStrictEqual([
             { 'severity': 'error', 'text': 'Error retrieving installed extensions: Could not read directory' }
         ]);
+    });
+
+    test('getInstalledExtensions() (extensions passed in)', async () => {
+        mockIsAppStudio.mockReturnValue(true);
+        const expectedResult = {
+            'vscode-ui5-language-assistant': { version: '3.3.0' },
+            'sap-ux-application-modeler-extension': { version: '1.7.4' },
+            'yeoman-ui': { version: '1.7.11' }
+        };
+        const logger = getLogger();
+        const extensions = [
+            {
+                id: 'SAPOS.yeoman-ui',
+                packageJSON: {
+                    name: 'yeoman-ui',
+                    version: '1.7.11'
+                }
+            },
+            {
+                id: 'SAPSE.sap-ux-application-modeler-extension',
+                packageJSON: {
+                    name: 'sap-ux-application-modeler-extension',
+                    version: '1.7.4'
+                }
+            },
+            {
+                id: 'SAPSE.vscode-ui5-language-assistant',
+                packageJSON: {
+                    name: 'vscode-ui5-language-assistant',
+                    version: '3.3.0'
+                }
+            }
+        ] as any as readonly Extension<any>[];
+
+        const result = await getInstalledExtensions(extensions, logger);
+        expect(result).toStrictEqual(expectedResult);
     });
 
     test('getCFCliToolVersion()', async () => {
@@ -102,9 +143,7 @@ describe('Test install functions', () => {
         const output = `some/path/to/lib/node_modules`;
         jest.spyOn(command, 'spawnCommand').mockResolvedValueOnce(output);
         jest.spyOn(fs, 'existsSync').mockReturnValue(true);
-        jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(
-            `{  "name": "@sap/generator-fiori",  "displayName": "SAP Fiori application",  "version": "1.7.5",  "description": "Create an SAPUI5 application using SAP Fiori elements or a freestyle approach"  }`
-        );
+        jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(mockPackageJson);
         const result = await getFioriGenVersion();
         expect(result).toStrictEqual(expectedResult);
     });
@@ -134,9 +173,7 @@ describe('Test install functions', () => {
         const output = `some/path/to/lib/node_modules`;
         jest.spyOn(command, 'spawnCommand').mockResolvedValueOnce(output);
         jest.spyOn(fs, 'existsSync').mockReturnValue(true);
-        jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(
-            `{  "name": "@sap/generator-fiori",  "displayName": "SAP Fiori application",  "version": "1.7.5",  "description": "Create an SAPUI5 application using SAP Fiori elements or a freestyle approach"  }`
-        );
+        jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(mockPackageJson);
         const result = await getFioriGenVersion();
         expect(result).toStrictEqual(expectedResult);
     });
@@ -151,9 +188,7 @@ describe('Test install functions', () => {
         jest.spyOn(command, 'spawnCommand').mockResolvedValueOnce(output);
 
         jest.spyOn(fs, 'existsSync').mockReturnValue(true);
-        jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(
-            `{  "name": "@sap/generator-fiori",  "displayName": "SAP Fiori application",  "version": "1.7.5",  "description": "Create an SAPUI5 application using SAP Fiori elements or a freestyle approach"  }`
-        );
+        jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(mockPackageJson);
         const result = await getFioriGenVersion();
         expect(result).toStrictEqual(expectedResult);
     });
@@ -164,6 +199,46 @@ describe('Test install functions', () => {
 
         const result = await getFioriGenVersion();
         expect(result).toStrictEqual('Not installed or not found');
+    });
+
+    test('getFioriGenVersion() (BAS) installed to NODE_PATH location', async () => {
+        // This test should not run on Windows, the code it tests is only run on BAS
+        // Only posix path.delimiter and path.separator are relevant
+        if (process.platform !== 'win32') {
+            mockIsAppStudio.mockReturnValue(true);
+            const expectedResult = '1.7.5';
+            const installedPath = `/installed/path/node_modules`;
+            const nodePathSaved = process.env.NODE_PATH;
+            process.env.NODE_PATH = `/some/global/node_modules:/another/global/node_modules:/yet/another/global/node_modules:${installedPath}`;
+
+            jest.spyOn(command, 'spawnCommand').mockResolvedValueOnce('/not/installed/path');
+            jest.spyOn(fs, 'existsSync').mockImplementation((filePath) => {
+                // Mocks finding the package at `installedPath`
+                if ((filePath as string).startsWith(installedPath)) {
+                    return true;
+                }
+                return false;
+            });
+            jest.spyOn(fs.promises, 'readFile').mockResolvedValueOnce(mockPackageJson);
+
+            let result = await getFioriGenVersion();
+            expect(result).toStrictEqual(expectedResult);
+
+            // Neg tests, NODE_PATH not set
+            process.env.NODE_PATH = '';
+            result = await getFioriGenVersion();
+            expect(result).toEqual(t('info.notInstalledOrNotFound'));
+
+            delete process.env.NODE_PATH;
+            result = await getFioriGenVersion();
+            expect(result).toEqual(t('info.notInstalledOrNotFound'));
+
+            if (nodePathSaved) {
+                process.env.NODE_PATH = nodePathSaved;
+            } else {
+                delete process.env.NODE_PATH;
+            }
+        }
     });
 
     test('getProcessVersions() (VSCODE)', async () => {
@@ -199,5 +274,35 @@ describe('Test install functions', () => {
         expect(messages).toStrictEqual([
             { 'severity': 'error', 'text': 'Error retrieving process versions: spawn error' }
         ]);
+    });
+
+    test('getInstalledExtensions() (VSCODE-Insiders)', async () => {
+        mockIsAppStudio.mockReturnValue(false);
+        process.env['VSCODE_IPC_HOOK'] = 'VSCode-main-insider';
+        const expectedResult = {
+            'yeoman-ui': { version: '2' },
+            'vscode-ui5-language-assistant': { version: '2' },
+            'xml-toolkit': { version: '2' }
+        };
+
+        const output = `SAPOS.yeoman-ui@2\nSAPOSS.vscode-ui5-language-assistant@2\nSAPOSS.xml-toolkit@2`;
+        jest.spyOn(command, 'spawnCommand').mockResolvedValueOnce(output);
+        const result = await getInstalledExtensions();
+        expect(result).toStrictEqual(expectedResult);
+    });
+
+    test('getInstalledExtensions() (VSCODE-Insiders 2)', async () => {
+        mockIsAppStudio.mockReturnValue(false);
+        process.env['TERM_PROGRAM_VERSION'] = '1.72.0-insider';
+        const expectedResult = {
+            'yeoman-ui': { version: '2' },
+            'vscode-ui5-language-assistant': { version: '2' },
+            'xml-toolkit': { version: '2' }
+        };
+
+        const output = `SAPOS.yeoman-ui@2\nSAPOSS.vscode-ui5-language-assistant@2\nSAPOSS.xml-toolkit@2`;
+        jest.spyOn(command, 'spawnCommand').mockResolvedValueOnce(output);
+        const result = await getInstalledExtensions();
+        expect(result).toStrictEqual(expectedResult);
     });
 });