+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.6.41
+
+### Patch Changes
+
+-   Updated dependencies [77ac998]
+    -   @sap-ux/axios-extension@1.0.0
+
+## 0.6.40
+
+### Patch Changes
+
+-   ed04f6f: chore(open-ux-tools) Upgrade Dev Dependencies and fix Audit issues
+-   Updated dependencies [ed04f6f]
+    -   @sap-ux/store@0.3.9
+    -   @sap-ux/axios-extension@0.14.1
+    -   @sap-ux/btp-utils@0.11.4
+    -   @sap-ux/logger@0.3.4
+
 ## 0.6.39
 
 ### Patch Changes