+++ packages/mockserver-config-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/mockserver-config-writer
 
+## 0.0.5
+
+### Patch Changes
+
+-   @sap-ux/ui5-config@0.16.3
+
+## 0.0.4
+
+### Patch Changes
+
+-   ed04f6f: chore(open-ux-tools) Upgrade Dev Dependencies and fix Audit issues
+-   Updated dependencies [ed04f6f]
+    -   @sap-ux/ui5-config@0.16.2
+
 ## 0.0.3
 
 ### Patch Changes