+++ packages/project-access/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/project-access
 
+## 1.1.1
+
+### Patch Changes
+
+-   @sap-ux/ui5-config@0.16.3
+
+## 1.1.0
+
+### Minor Changes
+
+-   470275c: Add option for findProjectRoot function to not throw errors when no match is found
+
+## 1.0.6
+
+### Patch Changes
+
+-   ed04f6f: chore(open-ux-tools) Upgrade Dev Dependencies and fix Audit issues
+-   Updated dependencies [ed04f6f]
+    -   @sap-ux/ui5-config@0.16.2
+
+## 1.0.5
+
+### Patch Changes
+
+-   8cc76b5: chore: consume App Descriptor (manifest.json) types from @ui5/manifest
+
 ## 1.0.4
 
 ### Patch Changes