+++ packages/ui5-config/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/ui5-config
 
+## 0.16.3
+
+### Patch Changes
+
+-   Updated dependencies [3327a3f]
+    -   @sap-ux/yaml@0.13.3
+
+## 0.16.2
+
+### Patch Changes
+
+-   ed04f6f: chore(open-ux-tools) Upgrade Dev Dependencies and fix Audit issues
+-   Updated dependencies [ed04f6f]
+    -   @sap-ux/yaml@0.13.2
+
 ## 0.16.1
 
 ### Patch Changes