+++ types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/types
 
+## 1.0.3
+
+### Patch Changes
+
+-   ed04f6f: chore(open-ux-tools) Upgrade Dev Dependencies and fix Audit issues
+
 ## 1.0.2
 
 ### Patch Changes